requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroplan==0.8
astropy==5.0.1
astropy-healpix==0.6
asttokens==2.0.5
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Comet==3.1.0
constantly==15.1.0
cryptography==36.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
future==0.18.2
h5py==3.7.0
healpy==1.15.2
hyperlink==21.0.0
idna==3.3
importlib-resources==5.8.0
incremental==21.3.0
iniconfig==1.1.1
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
iso8601==1.0.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.2
lalsuite==7.5
ligo-gracedb==2.7.7
ligo-segments==1.4.0
ligo.skymap==1.0.0
lscsoft-glue==3.0.1
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mysql-connector-python==8.0.29
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.3
notebook==6.4.12
numpy==1.22.3
orderedmultidict==1.0.1
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptemcee==1.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyerfa==2.0.0.1
Pygments==2.12.0
pymongo==4.1.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
pyRXP==3.0.1
pytest==7.1.2
python-dateutil==2.8.2
python-ligo-lw==1.8.0
pytz==2021.3
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
reproject==0.8
requests==2.27.1
scipy==1.8.1
Send2Trash==1.8.0
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.3.0
terminado==0.15.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
Twisted==22.2.0
typing_extensions==4.1.1
urllib3==1.26.8
voevent-parse==1.0.3
voeventdb.remote==1.1.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
xmltodict==0.12.0
zipp==3.8.0
zope.interface==5.4.0
gcn_kafka==0.3.0
